added snap-to-latest if scrolled to bottom
Added ref to chat-item container to track where whether it's been scrolled at all user is viewing the latest messages or reviewing older ones.

If scrolled to bottom, app stays snapped there and new chat-items emit an even AfterViewInit that pushes the chat upward

*** Solve Later ***
Scroll will not snap to bottom if chat-items are over 100px tall. There is a 100px buffer that determines if the user is scrolled to the bottom of the chat container. Verrrrrry long messages (over 100px tall) will break the method of checking this because the message height itself will consume the buffer and tell the snapToBottom function that the user isn't at the bottom and so don't scrolljack. ***